CCCP-227, fix: Resolve circular reference
fit into `eth-abi`'s breaking change
